Implement core::actions::Actions in OpenSSH.
This is NOT the full completion of the openssh module. Rather, it
specifically implements the four current Actions via OpenSSH. In other
words, if you specifically call the new methods, you can cause those
Actions to be sent to the client or executed on the client (depending on
the implementation flow).

Essentially, I'm working my way up from the bottom of the
controller-side code. sira-client is still NYI, but the openssh module
now tries to talk to sira-client.

Because serde to YAML is NYI, this commit does not actrually send the
Actions to sira-client. It sends placeholder text instead. It should be
trivial to replace that placeholder text with a YAML serialization of
the compiled Action when that code is implemented.