Core rust api (#2)
Implement and test the core Rust API, i.e. the code to instantiate, iterate over, and perform basic operations on the types in the core module. One notable exception is HostAction::compile, which is still NYI, because it makes more sense to handle this a bit later, once the precise details of domain knowledge are a bit clearer.