Add fields and docs to Action::Upload. WIP.
The first draft of the field names and documentation are complete.
However, the project does not compile in this state, as Serde is not yet
configured and the rest of the project is not yet updated.

This commit also modifies the documentation for Manifest to clarify how
relative task paths are handled. It adds a to-do item to check a related
corner case that I suspect contains a bug.