Refactor code to handle new fields. WIP.
I'm out of time today. This is as far as I have gotten on refactoring
the code base to reflect the updated definition of Action::Upload.

Bonus change: replace a wildcard `..` in an Action::LineInFile pattern
to improve future refactoring. This should be carried forth anywhere
else it makes sense.