Remove Action::Download. Explain local commands.
There are lots of good reasons not to implement either a Download action
or a LocalCommand action to run a command on the control node. This
commit removes the stub Download action, explains these reasons in
detail, and provides an example of how to accomplish a task that might
seem like it requires a Download or LocalCommand Action.