Rename Action::Shell to Action::Command (#9)
Originally, Action::Shell was intended to run full shell commands using the user's default shell. It proved infeasible to accomplish this without making assumptions about the shell. Instead, we spawn the requested processes directly. Therefore, the name Action::Shell is inaccurate.

This commit renames it to Action::Command and updates the docs to better explain the execution model.