Remove Task::user. Clear to-do: recursive upload.
Removing Task::user makes sense, now that Action is relatively mature.
There is neither any need nor any sensible definition for Task::user any
longer. Plus, the way to run a command as another user is clearly
documented in Action::Command.

Recursive upload is a tricky and substantial feature. There are no
current plans to implement it, but there is a partially implemented and
documented branch (as of this writing) for the feature. There is no need
to keep this as a to-do item.