Remove unneeded to-do.
I left the to-do when I was engrossed in another part of the project but
thought I might have found a bug. On closer examination, I was mistaken.

Specifically, the supposed bug involved how the function in question
handled manifest files in different folders, but the function only
accepts a single manifest file as an argument, so such a bug is
impossible. I manually verified this for extra peace of mind by testing
the following scenario, both with and without SSH key signing:

~/sira/a/manifest.yaml (includes b/task.yaml)
~/sira/a/b/task.yaml
~/sira/c/manifest.yaml (includes d/task.yaml)
~/sira/c/d/task.yaml

Running `sira a/manifest.yaml c/manifest.yaml` worked perfectly.