Add client-side implementation. WIP.
Appears to actually work, but clean-up, documentation, testing, etc. are
all incomplete.

Also, fix a security issue. When writing signature files in sira-client,
we called mktemp and captured its output as the signature file name.
However, we failed to trim the trailing newline from the output, so we
would actually write the signature to a different file. The symptom of
this issue was that temporary files were being left behind: fs::write
would create a new file, and fs:;remove_file would remove it, but
neither one was the original file.