Update permissions: o+rx (dirs), o+r (pub files).
This allows anyone on the system to leverage the allowed_signers and
public key files for any sorts of utility purposes. I wanted to do this
frequently in testing and found that the lack of read permission was as
frustrating as it was obviously unnecessary.

The action private key still has no permissions on other, of course.