Move tests for title() to correct location.
The tests for title() were in a findable but incorrect location due to
an oversight during development. title() moved around as it was
developed. This fixes an issue introduced in commit:
82517885fe1590f08695c3787a339c5d5d092773

The only effect of this commit is correcting internal test organization.
There are no functional changes to any end user programs.