Revise report() et al to simplify argument types.
Both calling code and internal code are now simpler. Plus, the type
checker and borrow checker are much happier with the simpler types, e.g.
when report() wants to reborrow stdout or stderr to pass to another
function. The new signature more clearly expresses the code's intent.

I believe the old signature was borrowed from a sound source somewhere
in std, but I'm not sure; it might have been intended to help MutexGuard
work in tests. Using `&mut *guard` with this new signature is better.