Change sira-client to accept file names.
WARNING: this change requires sira and sira-client to evolve at the same
time, as the previous commit's sira and this commit's sira-client are
not compatible! The auto-updater task on the README will not work. You
will need to transfer the new sira-client to all managed nodes. The
easiest way to do this may be to re-run sira-install for all nodes.

Alternatively, you may choose to do the following:
1. Install this version for your control user: cargo install ...
2. Copy the new sira-client binary (~/.cargo/bin/sira-client) elsewhere
3. Check out the previous commit: git checkout HEAD^
4. Install that version: cargo install ...
5. Copy the new binary to the control user's ~/.cargo/bin/sira-client
6. Use the auto-updater from the README
7. Check out main once again: git checkout main
8. Install the latest version for your control user: cargo install ...

Changes that break compatibility in this way are expected to be rare.
The compelling reason to do so in this case is that calling sira-client
correctly from the terminal was difficult or impossible, and this commit
makes calling it trivial.