Add [host] prefix to all lines of an action report.
This makes it much easier to filter the output log by host. It imposes
the slight trade-off of changing any dumped scripts or output more
significantly from the raw output, which means that someone using the
output directly needs to filter out the tags. This should be quite
simple to do, though, e.g. with regex.